requirements.txt
websockets>=12.0
aiohttp>=3.9
python-dotenv>=1.0
pyyaml>=6.0
sqlalchemy>=2.0
aiosqlite>=0.20
pandas>=2.0
numpy>=1.26
py-clob-client>=0.34.6
eth-account>=0.11
web3>=6.0
rich>=13.0
python-telegram-bot>=21.0
pytest>=8.0
pytest-asyncio>=0.23
